feat: Refactor folder structure and add FoodAdmin page
