+++ src/components/hourly/mockWeatherData.ts
