+++ src/pages/FoodAdmin.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const FoodAdmin = () => {
+  return <div>FoodAdmin</div>;
+};
+
+export default FoodAdmin;
